apps/cms-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyJWT==2.10.1
bcrypt==4.1.3
PyYAML==6.0.2
python-multipart==0.0.20
